+++ internal-packages/workflow-designer-ui/package.json
@@ -13,7 +13,7 @@
 	"dependencies": {
 		"@giselle-sdk/data-type": "workspace:*",
 		"@giselle-sdk/giselle-engine": "workspace:*",
-		"@giselle-sdk/language-models": "workspace:^",
+		"@giselle-sdk/language-model": "workspace:^",
 		"@giselle-sdk/text-editor": "workspace:^",
 		"@giselle/giselle-sdk-tsconfig": "workspace:*",
 		"@xyflow/react": "catalog:",
