+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/model/anthropic.tsx
@@ -1,7 +1,7 @@
 import {
 	AnthropicLanguageModel,
 	anthropicLanguageModels,
-} from "@giselle-sdk/language-models";
+} from "@giselle-sdk/language-model";
 import {
 	Select,
 	SelectContent,
