+++ packages/data-type/package.json
@@ -26,7 +26,7 @@
 		"vitest": "catalog:"
 	},
 	"dependencies": {
-		"@giselle-sdk/language-models": "workspace:^",
+		"@giselle-sdk/language-model": "workspace:^",
 		"ai": "catalog:",
 		"zod": "catalog:"
 	}
