+++ packages/giselle-engine/src/core/types.ts
@@ -1,4 +1,4 @@
-import type { LanguageModelProvider } from "@giselle-sdk/language-models";
+import type { LanguageModelProvider } from "@giselle-sdk/language-model";
 import type { Storage } from "unstorage";
 
 export interface GiselleEngineContext {
