+++ packages/giselle-sdk/package.json
@@ -36,7 +36,7 @@
 	},
 	"dependencies": {
 		"@giselle-sdk/giselle-engine": "workspace:^",
-		"@giselle-sdk/language-models": "workspace:^",
+		"@giselle-sdk/language-model": "workspace:^",
 		"@giselle-sdk/workspace": "workspace:^"
 	}
 }
