+++ packages/giselle-sdk/src/index.ts
@@ -1,2 +1,2 @@
 export * from "@giselle-sdk/giselle-engine/client";
-export type { LanguageModelProvider } from "@giselle-sdk/language-models";
+export type { LanguageModelProvider } from "@giselle-sdk/language-model";
